basketball-tracker-backend/requirements.txt
# Core
Flask==3.1.0
flask-cors==5.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
SQLAlchemy==2.0.36
alembic==1.14.0

# Database drivers
psycopg2-binary==2.9.10    # PostgreSQL (production)
# SQLite ships with Python — no extra driver needed for development

# Config / environment
python-dotenv==1.0.1

# Production server
gunicorn==23.0.0

# Dev / tooling
black==24.10.0
ruff==0.8.0
